refactoring CreditCardView .
